requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
singledispatch==3.4.0.3
tornado==4.3
wheel==0.24.0
WTForms==2.0.3.dev0
wtforms-tornado==0.0.2
PyYAML==3.11
Jinja2==2.8